requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
Django==3.2.20
django-enum==1.2.1
django-environ==0.10.0
django-filter==2.4.0
django-multiselectfield==0.1.12
django-password-validators==1.7.1
django-rest-framework==0.1.0
django-select-multiple-field==0.4.2
django-softdelete==0.11.4
django-utils==0.0.2
django-utils-six==2.0
django-utils2==3.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-utils==3.7.0
pytz==2023.3
pyzmq==25.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.14.3
typing_extensions==4.7.1
urllib3==2.0.2
wcwidth==0.2.6
